+++ smelter-api/src/output/mp4.rs
@@ -58,7 +58,7 @@ pub enum Mp4VideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval_ms: Option<u64>,
+        keyframe_interval_ms: Option<u32>,
     },
 }
