+++ smelter-api/src/output/mp4_into.rs
@@ -105,7 +105,8 @@ impl Mp4VideoEncoderOptions {
             } => core::VideoEncoderOptions::VulkanH264(core::VulkanH264EncoderOptions {
                 resolution: resolution.into(),
                 bitrate: bitrate.map(|bitrate| bitrate.try_into()).transpose()?,
-                keyframe_interval_ms: keyframe_interval_ms.unwrap_or(DEFAULT_KEYFRAME_INTERVAL),
+                keyframe_interval_ms: keyframe_interval_ms
+                    .unwrap_or(DEFAULT_KEYFRAME_INTERVAL as u32),
             }),
         };
         Ok(encoder_options)
