+++ smelter-api/src/output/rtp.rs
@@ -89,7 +89,7 @@ pub enum RtpVideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval_ms: Option<u64>,
+        keyframe_interval_ms: Option<u32>,
     },
 }
 #[derive(Debug, Serialize, Deserialize, Clone, JsonSchema)]
