+++ smelter-api/src/output/whip.rs
@@ -85,7 +85,7 @@ pub enum WhipVideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval_ms: Option<u64>,
+        keyframe_interval_ms: Option<u32>,
     },
     #[serde(rename = "any")]
     Any,
